+++ src/ducks/Watchlists/Widgets/Table/AssetsToggleColumns.module.scss
@@ -22,21 +22,6 @@
   }
 }
 
-.columnId {
-  display: flex;
-  width: 100%;
-  justify-content: center;
-}
-
-.assetCheckbox {
-  margin: auto;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  height: 100%;
-}
-
 .wrapper {
   min-width: 240px;
   padding: 12px 16px;
@@ -72,12 +57,6 @@
   margin-right: 6px;
 }
 
-.soon {
-  color: var(--jungle-green);
-  font-size: 12px;
-  margin-left: 5px;
-}
-
 .description {
   fill: var(--casper);
   position: relative;
@@ -86,29 +65,4 @@
 .tooltip {
   max-width: 400px;
   z-index: 1000;
-}
-
-.rank {
-  background-color: var(--porcelain);
-  color: var(--rhino);
-}
-
-.segments {
-  display: inline-flex;
-  width: fit-content;
-}
-
-.segment {
-  background-color: var(--porcelain);
-  color: var(--rhino);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  white-space: nowrap;
-
-  @include text('caption');
-}
-
-.segment + .segment {
-  margin-left: 4px;
 }
