+++ src/ducks/Watchlists/Widgets/Table/new-columns.module.scss
@@ -1,0 +1,41 @@
+@import "~@santiment-network/ui/mixins";
+
+.rank {
+  background-color: var(--porcelain);
+  color: var(--rhino);
+}
+
+.segments {
+  display: inline-flex;
+  width: fit-content;
+}
+
+.segment {
+  background-color: var(--porcelain);
+  color: var(--rhino);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  white-space: nowrap;
+
+  @include text('caption');
+}
+
+.segment + .segment {
+  margin-left: 4px;
+}
+
+.assetCheckbox {
+  margin: auto;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  height: 100%;
+}
+
+.columnId {
+  display: flex;
+  width: 100%;
+  justify-content: center;
+}
